move to Windows directory
